PIMS-1281 Receive Order > NumberFormatException: For input string: "ff808181352fd0a401352fde19360001"
git-svn-id: https://dev.pih-emr.org/svn/repository/projects/warehouse/trunk@11137 f0ac6d2b-dfc5-4988-9158-fc266d298d87

+++ grails-app/controllers/org/pih/warehouse/order/ReceiveOrderWorkflowController.groovy
@@ -25,7 +25,7 @@ class ReceiveOrderWorkflowController {
 				// create a new shipment instance if we don't have one already
 				def orderCommand = new OrderCommand();
 				if (params.id) {
-					orderCommand = orderService.getOrder(params.id as int, session.user.id as int)
+					orderCommand = orderService.getOrder(params.id, session.user.id as int)
 				}
 				else {
 					flash.message = "${warehouse.message(code: 'default.not.found.message', args: [warehouse.message(code: 'order.label', default: 'Order'), params.id])}"
